adapt to new API and remove translate arg
